canbenli/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==4.0.3
django-admin-thumbnails==0.2.6
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-currencies==0.10.1
django-js-asset==2.0.0
django-mptt==0.13.4
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
jsonfield==2.1.1
Pillow==9.0.1
platformdirs==2.5.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.13.1
virtualenvwrapper-win==1.2.7
